+++ app/ui-admin/client/sidebarItems.js
@@ -1,6 +1,8 @@
 import { ReactiveVar } from 'meteor/reactive-var';
 import { Tracker } from 'meteor/tracker';
 
+import { hasPermission } from '../../authorization/client';
+
 const items = new ReactiveVar([]);
 
 export const registerAdminSidebarItem = (itemOptions) => {
@@ -8,3 +10,10 @@ export const registerAdminSidebarItem = (itemOptions) => {
 };
 
 export const getSidebarItems = () => items.get().filter((option) => !option.permissionGranted || option.permissionGranted());
+
+registerAdminSidebarItem({
+	href: 'invites',
+	i18nLabel: 'Invites',
+	icon: 'user-plus',
+	permissionGranted: () => hasPermission('create-invite-links'),
+});
